% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVariables.R
\name{editingBP}
\alias{editingBP}
\title{Editing final blood pressure file}
\usage{
editingBP(crea.dataset = crea.rep, bptype_data = bpfile, BPtype = c())
}
\arguments{
\item{crea.dataset}{dataset with creatinine ReadCodes, at least three creatinine measurements per patient.}

\item{bptype_data}{blood pressure data gotten from addBP function.}

\item{BPtype}{a character string. E.g. "diastolic" or "systolic". How ever you type it, it will be converter to lower case. This will then be pasted with "CodeValue" and "Flag"}
}
\value{
it returns a edited diastolic BP or systolic BP data frame with CodeValue column with BP value and Flag column with TRUE or FALSE. If TRUE, the BP data for this row is older than 30 days.
}
\description{
This function will edit column names of the file gotten from addBP(). It will append column names from crea.rep dataset first, and then customly add the names
of the two new columns which are last columns (19 and 20 respectfully). The columns' names will have a name diastolicCodeValue or systolicCodeValue and
diastolicFag or systolicFlag, depending on which file do you analyse. It will also add the formerge column where it
creates a identifier for merging. This column is created by simply pasting PatientID, event.date and CodeValue of the bp file.
}
\examples{
\dontrun{
editingBP(crea.rep, crea_dia.txt, BPtype="diastolic")
}

}
