% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVariables.R
\name{mergeMe}
\alias{mergeMe}
\title{Merging it all together}
\usage{
mergeMe(df1, df2, colsToSelect = NULL)
}
\arguments{
\item{df1}{one of the datasets from variableDataEditing() function.}

\item{df2}{one of the datasets from variableDataEditing() function.}

\item{colsToSelect}{Default is NULL. A character vector of columns which you want to save from df2. No quotation marks. If you leave it as NULL, only formerge column will be added to the final data frame.}
}
\description{
Helps to merge four datasets (BMI, Systolic BP, Diastolic BP and Diabetes). First one you submit (df1) will be the "main", meaning that all the columns
of this data set will be preserved. In the second dataset only some columns will be preserved (according to your demands). You customize that
in the colsToSelect argument. "formerge" column is always preserved, and you specify the others.
}
\examples{
\dontrun{
mergeMe(bmidata, sysdata, colsToSelect=c(systolicCodeValue, systolicFlag))
mergeMe(bmisys, diastdata, colsToSelect=c(diastolicCodeValue, diastolicFlag))
mergeMe(bmisysdiast, diabetes.data, colsToSelect=c("Diabetes","diabetes.exposure"))
}

}
