% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVariables.R
\name{variableDataEditing}
\alias{variableDataEditing}
\title{Editing variable files from add*() functions}
\usage{
variableDataEditing(crea.dataset = crea.rep, variable.data = variable.data,
  BPtype = NULL, toedit = c("BP", "BMI", "diabetes"))
}
\arguments{
\item{crea.dataset}{dataset with creatinine ReadCodes, at least three creatinine measurements per patient.}

\item{variable.data}{bmi, diabetes or blood pressure data gotten from one of the add*() functions in the package.}

\item{BPtype}{a character string. E.g. "diastolic" or "systolic". How ever you type it, it will be converter to lower case. Default is NULL}

\item{toedit}{A vector. Can be "BP", "BMI" or "diabetes". This will tell the function which file editing you want to do.}
}
\value{
it returns a edited BMI, diaebete, diastolic BP or systolic BP data frame.
}
\description{
This function will edit column names of the file gotten from add*() function. It will append column names from crea.rep dataset first, and then customly add the names
of the new column(s), depending on which file are you editing (BP, BMI or diabetes).
It will also add the "formerge" column where it creates a identifier for merging. This column is created by simply pasting PatientID, event.date and CodeValue of the variable file.
}
\details{
If toedit="BP", you have to specify BPtype ("diastolic" or "systolic"). This will create a edited BP dataframe - it renames last two columns to BP code value and "Flag". Regardless
of what you chose in the "toedit", function will also add the "formerge" column created for the purposes of merging all the variables to the final file.
I suggest saving the final files as RDS (saveRDS(editeddiastolicfile, "path/to/file.rds")), as it uses less space and loads faster.
}
\examples{
\dontrun{
editeddiastolicfile <- variableDataEditing(crea.rep, crea_dia.txt, BPtype="diastolic", toedit="BP")
editedBMIfile <- variableDataEditing(crea.rep, crea_dia.txt, toedit="BMI")
}

}
