% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVariables.R
\name{addBP}
\alias{addBP}
\title{Adding blood pressure  data to your final file}
\usage{
addBP(crea.dataset = crea.rep, bpdata = bpdata, BPReadCode, filename)
}
\arguments{
\item{crea.dataset}{dataset with creatinine ReadCodes, at least three creatinine measurements per patient.}

\item{bpdata}{blood pressure data filtered out from SIR data or similar. It has to be clean data, no NA, no duplicates. Has to contain ReadCodes for blood pressure type, event.date of blood pressure measurement and patient ID.}

\item{BPReadCode}{blood pressure ReadCode, character vector of length one (see examples).}

\item{filename}{name of the final .txt file your output will be saved to. Can be a path also. Should be a txt file because of write.table function inside.}
}
\value{
this function returns out file of lapply function. Can be useful to save the output to a file in case an error occurrs. You can find error details in that output file. The creatinine data with diabetes is automatically saved, row by row.
}
\description{
This function takes your dataset (crea.rep or similar) and finds blood pressure information on each patient in crea dataset.
It is up to you which ReadCode you want to analyse. You can only analyse full ReadCode (not the ones that only begin with some string).
This function also adds Flag column to the final data set. If the BP measure is more than 30 days old, the Flag column for this patient's event.date will have TRUE.
}
\details{
it uses lapply, not mclapply.

if the dplyr package is not installed already, it will be.
}
\examples{
\dontrun{
addBMI(crea.dataset=crea.rep, bpdata, BPReadCode=c("246A."), filename="creaWithDiabetes.txt")
}

}
