% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVariables.R
\name{addDiabetes}
\alias{addDiabetes}
\title{Adding diabetes data to your final file}
\usage{
addDiabetes(crea.dataset, diabetes.dataset, NbOfCores = 4L, filename)
}
\arguments{
\item{crea.dataset}{dataset with creatinine ReadCodes, at least three creatinine measurements per patient.}

\item{diabetes.dataset}{file with extracted BMI codes from SIR data (codes starting with" 22K"); duplicates removed, NA free}

\item{NbOfCores}{how much cores do you want to use for your parallel process? Will be added to getOption() from mclapply. Default is 4.}

\item{filename}{name of the final .txt file your output will be saved to. Can be a path also. Should be a txt file because of write.table function inside.}
}
\value{
This function returns out file of mclapply function. Can be useful to save the output to a file in case an error occurrs. You can find error details in that output file. The creatinine data with diabetes is automatically saved to a .txt file, row by row, and can later be accessed by using "read.table" function.
}
\description{
This function takes your dataset (crea.rep or similar) and finds diabetes information on each patient in crea dataset. For each patient
it first finds the earliest event.date of diabetes and then searches for the closest creatinine event.date (making sure that crea event.date
was chronologically *after* the diabetes date)
}
\details{
This function uses parallel lapply - mclapply from "parallel" package. It's recommended to either run it in the background of your PC or on the cluster. The column used for time analysis is named "event.date".
}
\examples{
\dontrun{
addBMI(crea.dataset, diabetes.dataset = diabetes, NbOfCores=4L, filename = "creaWithDiabetes.txt")
}

}
